add movie methods implementation

+++ src/main/java/com/watchworthy/api/dto/MovieDTO.java
@@ -1,9 +1,6 @@
 package com.watchworthy.api.dto;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.time.LocalDate;
 
@@ -13,7 +10,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class MovieDTO {
 
-    private Long id;
+    private Integer id;
     private String title;
     private String overview;
     private LocalDate releaseDate;
